pyproject.toml
[project]
name = "pi-card"
version = "0.1.0"
description = "Raspberry Pi AI voice assistant"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pi_card"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
